:hourglass: Support timer rule deletion

+++ src/pvr_client/timer.cpp
@@ -257,61 +257,32 @@ PVR_ERROR AddTimer(const PVR_TIMER& timer)
 
 PVR_ERROR DeleteTimer(const PVR_TIMER& timer, bool bForceDelete)
 {
-    if (timer.iTimerType == TIMER_MANUAL_RESERVED || timer.iTimerType == TIMER_PATTERN_MATCHED) {
+    switch (timer.iTimerType) {
+    case TIMER_MANUAL_RESERVED: {
         for (const auto program : g_schedule.programs) {
-            if ((int)program.channelId != timer.iClientChannelUid) {
-                continue;
-            }
-            if (program.startAt == timer.startTime && program.endAt == timer.endTime) {
-                time_t now;
-                time(&now);
-                if (program.startAt < now && now < program.endAt) { // Ongoing recording
-                    if (epgstation::api::deleteReserves(std::to_string(program.id)) != epgstation::api::REQUEST_FAILED) { // Cancel recording
-                        XBMC->Log(ADDON::LOG_NOTICE, "Cancel ongoing recording program: %s", std::to_string(program.id).c_str());
-                        sleep(5);
-                        PVR->TriggerRecordingUpdate();
-                        PVR->TriggerTimerUpdate();
-                        return PVR_ERROR_NO_ERROR;
-                    } else {
-                        XBMC->Log(ADDON::LOG_ERROR, "Failed to cancel recording program: %s", std::to_string(program.id).c_str());
-                        XBMC->QueueNotification(ADDON::QUEUE_ERROR, "Failed to cancel recording program: %s", std::to_string(program.id).c_str());
-                        return PVR_ERROR_SERVER_ERROR;
-                    }
-                } else if (timer.iTimerType == TIMER_MANUAL_RESERVED) {
-                    if (epgstation::api::deleteReserves(std::to_string(program.id)) != epgstation::api::REQUEST_FAILED) {
-                        XBMC->Log(ADDON::LOG_NOTICE, "Delete manual reserved program: %s", std::to_string(program.id).c_str());
-                        sleep(1);
-                        PVR->TriggerRecordingUpdate();
-                        PVR->TriggerTimerUpdate();
-                        return PVR_ERROR_NO_ERROR;
-                    } else {
-                        XBMC->Log(ADDON::LOG_ERROR, "Failed to delete reserved program: %s", std::to_string(program.id).c_str());
-                        XBMC->QueueNotification(ADDON::QUEUE_ERROR, "Failed to delete reserved program: %s", std::to_string(program.id).c_str());
-                        return PVR_ERROR_SERVER_ERROR;
-                    }
+            if ((int)program.channelId == timer.iClientChannelUid && program.startAt == timer.startTime && program.endAt == timer.endTime) {
+                if (epgstation::api::deleteReserves(timer.strDirectory) != epgstation::api::REQUEST_FAILED) {
+                    XBMC->Log(ADDON::LOG_NOTICE, "Delete reserved program: %s", timer.strDirectory);
+                    sleep(1);
+                    PVR->TriggerRecordingUpdate();
+                    PVR->TriggerTimerUpdate();
+                    return PVR_ERROR_NO_ERROR;
                 } else {
-                    if (epgstation::api::deleteReserves(timer.strDirectory) != epgstation::api::REQUEST_FAILED) {
-                        XBMC->Log(ADDON::LOG_NOTICE, "Skip reserving: %s", timer.strDirectory);
-                        sleep(1);
-                        PVR->TriggerRecordingUpdate();
-                        PVR->TriggerTimerUpdate();
-                        return PVR_ERROR_NO_ERROR;
-                    } else {
-                        XBMC->Log(ADDON::LOG_ERROR, "Failed to skip reserved program: %s", std::to_string(program.id).c_str());
-                        XBMC->QueueNotification(ADDON::QUEUE_ERROR, "Failed to skip reserved program: %s", std::to_string(program.id).c_str());
-                        return PVR_ERROR_SERVER_ERROR;
-                    }
+                    XBMC->Log(ADDON::LOG_ERROR, "Failed to delete reserved program: %s", std::to_string(program.id).c_str());
+                    XBMC->QueueNotification(ADDON::QUEUE_ERROR, "Failed to delete reserved program: %s", std::to_string(program.id).c_str());
+                    return PVR_ERROR_SERVER_ERROR;
                 }
             }
         }
-
         XBMC->Log(ADDON::LOG_ERROR, "Failed to delete timer: nothing matched");
         XBMC->QueueNotification(ADDON::QUEUE_ERROR, "Failed to delete timer: nothing matched");
         return PVR_ERROR_FAILED;
-    } else {
+    }
+    default: {
         XBMC->Log(ADDON::LOG_ERROR, "Unknown timer type for deletion request: %d", timer.iTimerType);
         XBMC->QueueNotification(ADDON::QUEUE_ERROR, "Unknown timer type for deletion request: %d", timer.iTimerType);
-        return PVR_ERROR_FAILED;
+        return PVR_ERROR_NOT_IMPLEMENTED;
+    }
     }
 }
